Перенес вызов функции создания коллекции из страницы в шапку, сделал необязательными некоторые свойства
