+++ src/blocks/create-empty/header/create-empty-header.js
@@ -1,10 +1,15 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { NavLink } from 'react-router-dom';
+import { NavLink, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { Icon } from '../../../blocks';
 
 import './create-empty-header.scss';
+
+import { cardBlue } from '../../../variables.scss';
+
+import { createCollection } from '../../../reducers/create-collection.reducer';
+
 
 class CreateEmptyHeader extends Component {
     constructor() {
@@ -26,9 +31,9 @@ class CreateEmptyHeader extends Component {
         } = this.props;
         const { title, description, hashTags } = nextProps;
 
-        if ((currTitle !== title) ||
-            (currDescription !== description) ||
-            (currHashTags !== hashTags)
+        if ((title && currTitle !== title) ||
+            (description && currDescription !== description) ||
+            (hashTags && currHashTags !== hashTags)
         ) {
             this.setSubmitStatus(nextProps);
         }
@@ -44,9 +49,32 @@ class CreateEmptyHeader extends Component {
         });
     }
 
-    render() {
-        const { submitCallback } = this.props;
+    changeRoute = () => {
+        this.props.history.push({ pathname: './feed' });
+    };
 
+    hexToRGB = (hex) => {
+        const r = parseInt(hex.slice(1, 3), 16);
+        const g = parseInt(hex.slice(3, 5), 16);
+        const b = parseInt(hex.slice(5, 7), 16);
+
+        return `rgb(${r}, ${g}, ${b})`;
+    };
+
+    handleSubmitData = () => {
+        const body = {
+            description: this.props.description,
+            name: this.props.title,
+            photo: 'https://pp.userapi.com/c543100/v543100915/2fbfb/IoVG_UEW-yw.jpg',
+            color: this.hexToRGB(cardBlue),
+            tags: ['59a7e38c7db98b35471fed6d', '59a7e38c7db98b35471fed67'],
+        };
+
+        this.props.createCollection(body, this.props.token, this.changeRoute);
+    };
+
+
+    render() {
         return (
             <header className="create-empty-header">
                 <NavLink to={'/feed'}>
@@ -55,18 +83,26 @@ class CreateEmptyHeader extends Component {
                 <h4 className="create-empty-header__title">Новая тема</h4>
                 <button
                     className={`create-empty-header__submit ${this.state.submitStatus ? 'create-empty-header__submit--active' : ''}`}
-                    onClick={submitCallback}
+                    onClick={this.handleSubmitData}
                 >Создать</button>
             </header>
         );
     }
 }
 
+CreateEmptyHeader.defaultProps = {
+    title: '',
+    description: '',
+    hashTags: [],
+};
+
 CreateEmptyHeader.propTypes = {
-    submitCallback: PropTypes.func.isRequired,
-    title: PropTypes.string.isRequired,
-    description: PropTypes.string.isRequired,
-    hashTags: PropTypes.array.isRequired,
+    title: PropTypes.string,
+    description: PropTypes.string,
+    hashTags: PropTypes.array,
+    token: PropTypes.string.isRequired,
+    createCollection: PropTypes.func.isRequired,
+    history: PropTypes.any.isRequired,
 };
 
 export default connect(
@@ -74,5 +110,7 @@ export default connect(
         title: state.createCollection.title,
         description: state.createCollection.description,
         hashTags: state.createCollection.hashTags,
+        token: state.app.token,
     }),
-)(CreateEmptyHeader);
+    { createCollection },
+)(withRouter(CreateEmptyHeader));
